requirements.txt
spacy>=3.7.2
python-docx>=1.0.1
loguru>=0.7.2 